refactor(wire): split wire.rs into per-domain modules
src/wire.rs — the single source of truth for the serialized
JSON/YAML/TOML/CBOR shapes — held 25 wire structs plus their
From<&Stats> conversions and two test modules at 1360 production SLOC,
over the 800-line self-scan limit. Split the per-domain struct groups
into a wire/ subdirectory:

  - wire/metrics.rs  the 14 per-metric wire structs (Abc, Cognitive,
                     Halstead, Loc, ...) each with its From<&<m>::Stats>
  - wire/vcs.rs      the VCS wire structs (Vcs, VcsReport*, VcsTrend*)

Pure code movement: bodies verbatim, no logic or serialized-shape
change. wire.rs keeps the aggregate shapes (CodeMetrics, FuncSpace,
Ops, FunctionSpan), the helpers, the non_finite module, and both test
modules. The structs are public API (`pub mod wire;`;
`serde_json::from_str::<wire::FuncSpace>`), so they are re-exported with
`pub use metrics::*` / `pub use vcs::*`, preserving every `wire::<Struct>`
path. lib.rs is byte-identical. The vcs arm is entirely
`#[cfg(feature = "vcs-git")]`-gated, so its `mod`/`pub use` carry that
cfg (with a why-comment) to keep default-feature builds warning-free —
no shape or gating change.

wire.rs 1360 -> 486 production SLOC; wire/metrics.rs 573, wire/vcs.rs 317.
